added part2_task1
